backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1
python-telegram-bot>=20.7
websockets>=12.0
Pillow>=10.2.0
httpx>=0.26.0
aiofiles>=23.2.1
python-multipart>=0.0.6
streamlit>=1.31.0
requests>=2.31.0
google-genai>=1.0.0
groq>=0.4.0
